Make 'defaultResourceNamePattern' configurable for both Warn and Fix plugins
### What's done:
* Added warn and fix config - resourceNamePattern
