+++ save-common/src/commonMain/kotlin/org/cqfn/save/core/plugin/Plugin.kt
@@ -18,9 +18,9 @@ abstract class Plugin(open val testConfig: TestConfig, private val testFiles: Li
      *
      * @return a sequence of [TestResult]s for each group of test resources
      */
-    open fun execute(): Sequence<TestResult> = handleFiles(
+    fun execute(): Sequence<TestResult> = handleFiles(
         // todo: pass individual groups of files to handleFiles? Or it will play bad with batch mode?
-        discoverTestFiles(testConfig.directory, null)
+        discoverTestFiles(testConfig.directory)
     )
 
     /**
@@ -36,11 +36,10 @@ abstract class Plugin(open val testConfig: TestConfig, private val testFiles: Li
      * for execution of individual tests.
      *
      * @param root root [Path], from where discovering should be started
-     * @param regex default regex for actual warnings in the tool output
      * @return a sequence of files, grouped by test
      */
-    fun discoverTestFiles(root: Path, regex: Regex?): Sequence<List<Path>> {
-        val rawTestFiles = rawDiscoverTestFiles(root.resourceDirectories(), regex)
+    fun discoverTestFiles(root: Path): Sequence<List<Path>> {
+        val rawTestFiles = rawDiscoverTestFiles(root.resourceDirectories())
         return if (testFiles.isNotEmpty()) {
             rawTestFiles.filter { paths ->
                 // test can be specified by the name of one of it's files
@@ -55,10 +54,9 @@ abstract class Plugin(open val testConfig: TestConfig, private val testFiles: Li
      * Discover groups of resource files which will be used to run tests.
      *
      * @param resourceDirectories a sequence of [Path]s, which contain this plugin's resources
-     * @param regex default regex for actual warnings in the tool output
      * @return a sequence of files, grouped by test
      */
-    abstract fun rawDiscoverTestFiles(resourceDirectories: Sequence<Path>, regex: Regex?): Sequence<List<Path>>
+    abstract fun rawDiscoverTestFiles(resourceDirectories: Sequence<Path>): Sequence<List<Path>>
 
     /**
      * Returns a sequence of directories, where resources for this plugin may be located.
