+++ save-plugins/fix-plugin/src/commonMain/kotlin/org/cqfn/save/plugins/fix/FixPlugin.kt
@@ -34,14 +34,6 @@ class FixPlugin(testConfig: TestConfig, testFiles: List<String> = emptyList()) :
         .newTag { start -> if (start) "<" else ">" }
         .build()
 
-    override fun execute(): Sequence<TestResult> {
-        val fixPluginConfig = testConfig.pluginConfigs.filterIsInstance<FixPluginConfig>().single()
-        val regex = fixPluginConfig.resourceNamePattern ?: defaultResourceNamePattern
-        return handleFiles(
-            discoverTestFiles(testConfig.directory, regex)
-        )
-    }
-
     override fun handleFiles(files: Sequence<List<Path>>): Sequence<TestResult> {
         val fixPluginConfig = testConfig.pluginConfigs.filterIsInstance<FixPluginConfig>().single()
         logInfo("Discovered the following file pairs for comparison: $files")
@@ -69,24 +61,28 @@ class FixPlugin(testConfig: TestConfig, testFiles: List<String> = emptyList()) :
             }
     }
 
-    override fun rawDiscoverTestFiles(resourceDirectories: Sequence<Path>, regex: Regex?): Sequence<List<Path>> = resourceDirectories
-        .map { FileSystem.SYSTEM.list(it) }
-        .flatMap { files ->
-            files.groupBy {
-                val matchResult = (regex ?: defaultResourceNamePattern).matchEntire(it.name)
-                matchResult?.groupValues?.get(1)  // this is a capture group for the start of file name
-            }
-                .filter { it.value.size > 1 && it.key != null }
-                .mapValues { (name, group) ->
-                    require(group.size == 2) { "Files should be grouped in pairs, but for name $name these files have been discovered: $group" }
-                    listOf(
-                        group.first { it.name.contains("Expected.") },
-                        group.first { it.name.contains("Test.") }
-                    )
+    override fun rawDiscoverTestFiles(resourceDirectories: Sequence<Path>): Sequence<List<Path>> {
+        val fixPluginConfig = testConfig.pluginConfigs.filterIsInstance<FixPluginConfig>().single()
+        val regex = fixPluginConfig.resourceNamePattern ?: defaultResourceNamePattern
+        return resourceDirectories
+            .map { FileSystem.SYSTEM.list(it) }
+            .flatMap { files ->
+                files.groupBy {
+                    val matchResult = (regex).matchEntire(it.name)
+                    matchResult?.groupValues?.get(1)  // this is a capture group for the start of file name
                 }
-                .values
-        }
-        .filter { it.isNotEmpty() }
+                    .filter { it.value.size > 1 && it.key != null }
+                    .mapValues { (name, group) ->
+                        require(group.size == 2) { "Files should be grouped in pairs, but for name $name these files have been discovered: $group" }
+                        listOf(
+                            group.first { it.name.contains("Expected.") },
+                            group.first { it.name.contains("Test.") }
+                        )
+                    }
+                    .values
+            }
+            .filter { it.isNotEmpty() }
+    }
 
     private fun Patch<String>.formatToString() = deltas.joinToString("\n") { delta ->
         when (delta) {
