+++ save-plugins/fix-plugin/src/commonMain/kotlin/org/cqfn/save/plugins/fix/FixPluginConfig.kt
@@ -4,19 +4,12 @@
 package org.cqfn.save.plugins.fix
 
 import org.cqfn.save.core.plugin.PluginConfig
+import org.cqfn.save.core.plugin.RegexSerializer
 
 import okio.Path
 
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.Serializer
 import kotlinx.serialization.UseSerializers
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 
 /**
  * Some fields by default are null, instead of some natural value, because of the fact, that in stage of merging
@@ -25,7 +18,7 @@ import kotlinx.serialization.encoding.Encoder
  *
  * @property execCmd a command that will be executed to mutate test file contents
  * @property destinationFileSuffix [execCmd] should append this suffix to the file name after mutating it.
- * @property resourceNamePattern
+ * @property resourceNamePattern regex for the name of the test files
  */
 @Serializable
 data class FixPluginConfig(
@@ -64,17 +57,4 @@ data class FixPluginConfig(
     companion object {
         internal val defaultResourceNamePattern = Regex("""(.+)(Expected|Test)\.[\w\d]+""")
     }
-}
-
-@OptIn(ExperimentalSerializationApi::class)
-@Serializer(forClass = Regex::class)
-object RegexSerializer : KSerializer<Regex> {
-    override val descriptor: SerialDescriptor =
-            PrimitiveSerialDescriptor("regex", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: Regex) {
-        encoder.encodeString(value.pattern)
-    }
-
-    override fun deserialize(decoder: Decoder): Regex = Regex(decoder.decodeString())
 }
