+++ save-plugins/fix-plugin/src/commonTest/kotlin/org/cqfn/save/plugins/fix/FixPluginTest.kt
@@ -32,13 +32,8 @@ class FixPluginTest {
         fs.createFile(tmpDir / "Test1Test.java")
         fs.createFile(tmpDir / "Test1Expected.java")
 
-        val catCmd = if (isCurrentOsWindows()) "type" else "cat"
-        val config = FixPluginConfig(
-            "$catCmd ${tmpDir / "resource"}"
-        )
-
-        val pairs = FixPlugin(TestConfig(tmpDir / "Test1Test.java", null, mutableListOf()))
-            .discoverTestFiles(tmpDir, config.resourceNamePattern)
+        val pairs = FixPlugin(TestConfig(tmpDir / "Test1Test.java", null, mutableListOf(FixPluginConfig(""))))
+            .discoverTestFiles(tmpDir)
             .map { it.first() to it.last() }
             .toList()
 
@@ -59,13 +54,8 @@ class FixPluginTest {
         fs.createFile(tmpDir / "NowCompletelyDifferentExpected.java")
         fs.createFile(tmpDir / "AndNowCompletelyDifferent.java")
 
-        val catCmd = if (isCurrentOsWindows()) "type" else "cat"
-        val config = FixPluginConfig(
-            "$catCmd ${tmpDir / "resource"}"
-        )
-
-        val pairs = FixPlugin(TestConfig(tmpDir / "Something.java", null, mutableListOf()))
-            .discoverTestFiles(tmpDir, config.resourceNamePattern)
+        val pairs = FixPlugin(TestConfig(tmpDir / "Something.java", null, mutableListOf(FixPluginConfig(""))))
+            .discoverTestFiles(tmpDir)
             .map { it.first() to it.last() }
             .toList()
