+++ save-plugins/warn-plugin/src/commonMain/kotlin/org/cqfn/save/plugin/warn/WarnPlugin.kt
@@ -29,29 +29,24 @@ class WarnPlugin(testConfig: TestConfig, testFiles: List<String> = emptyList()) 
     private val fs = FileSystem.SYSTEM
     private val pb = ProcessBuilder()
 
-    override fun execute(): Sequence<TestResult> {
-        val warnPluginConfig = testConfig.pluginConfigs.filterIsInstance<WarnPluginConfig>().single()
-        val regex = warnPluginConfig.resourceNamePattern ?: defaultResourceNamePattern
-        return handleFiles(
-            discoverTestFiles(testConfig.directory, regex)
-        )
-    }
-
     override fun handleFiles(files: Sequence<List<Path>>): Sequence<TestResult> {
         val warnPluginConfig = testConfig.pluginConfigs.filterIsInstance<WarnPluginConfig>().single()
         val generalConfig = testConfig.pluginConfigs.filterIsInstance<GeneralConfig>().singleOrNull()
-        val regex = warnPluginConfig.resourceNamePattern ?: defaultResourceNamePattern
-        return discoverTestFiles(testConfig.directory, regex).map { resources ->
+        return discoverTestFiles(testConfig.directory).map { resources ->
             handleTestFile(resources.single(), warnPluginConfig, generalConfig)
         }
     }
 
-    override fun rawDiscoverTestFiles(resourceDirectories: Sequence<Path>, regex: Regex?): Sequence<List<Path>> = resourceDirectories
-        .map { directory ->
-            FileSystem.SYSTEM.list(directory)
-                .filter { (regex ?: defaultResourceNamePattern).matches(it.name) }
-        }
-        .filter { it.isNotEmpty() }
+    override fun rawDiscoverTestFiles(resourceDirectories: Sequence<Path>): Sequence<List<Path>> {
+        val warnPluginConfig = testConfig.pluginConfigs.filterIsInstance<WarnPluginConfig>().single()
+        val regex = warnPluginConfig.resourceNamePattern ?: defaultResourceNamePattern
+        return resourceDirectories
+            .map { directory ->
+                FileSystem.SYSTEM.list(directory)
+                    .filter { (regex ?: defaultResourceNamePattern).matches(it.name) }
+            }
+            .filter { it.isNotEmpty() }
+    }
 
     @Suppress(
         "TOO_LONG_FUNCTION",
